wrap handleSelectCompany in useCallback, change useSetCompanyHandler to .ts file, create clear variable names in handleSetCompany function
